+++ src/Xteq5UserInterface/Xteq5UIConstant.cs
@@ -4,6 +4,7 @@ using System.IO;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using Yamua;
 
 namespace Xteq5
 {
@@ -23,13 +24,28 @@ namespace Xteq5
             }
         }
 
+        /// <summary>
+        /// Returns the content of licenses\license.txt or the error text if the file was not found
+        /// </summary>
+        public static string LicenseTxtContent
+        {
+            get
+            {
+                string filePath = ApplicationInformation.Instance.ParentFolder + @"licenses\license.txt";
+
+                return File.Exists(filePath) ? File.ReadAllText(filePath, Encoding.Default) : "Unable to load " + filePath;
+            }
+        }
+
         //MTH: These are static readonly fields BY PURPOSE (not CONST) to avoid that the compiler performs an copy+paste to the assembly of the consumer.
         //See [Const Strings - a very convenient way to shoot yourself in the foot](http://www.stum.de/2009/01/14/const-strings-a-very-convenient-way-to-shoot-yourself-in-the-foot/) by [Michael Stum](http://www.stum.de/)        
 
 
-        public static readonly string ProjectURL = "https://github.com/texhex/Xteq5/wiki/";
+        public static readonly string DocumentationURL = "https://github.com/texhex/Xteq5/wiki/";
 
         public static readonly string HomepageURL = "http://www.xteq5.com/";
+
+        public static readonly string HowtoUseURL = "https://github.com/texhex/Xteq5/wiki/How-to-use";
 
 
     }